lekce10/interview/requirements.txt
# Core web dependencies
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
gunicorn>=21.2.0
python-multipart>=0.0.9
starlette>=0.36.3

# Twilio
twilio==8.12.0

# OpenAI and LangChain
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langgraph>=0.0.18

# Azure
azure-functions>=1.17.0

# Utils
python-dotenv>=0.19.0
typing-extensions>=4.5.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

